Stat change
Changed attack rate to be float instead of double. With double, 0.4
atkRate would show up as 0.400000005964

+++ Assets/Scripts/AllyUnitStats.cs
@@ -14,7 +14,7 @@ public class AllyUnitStats : MonoBehaviour {
     public int currentHealth;
     public int attackDmg;
     public int attackRange;
-    public double attackRate;
+    public float attackRate;
     public float energy;
     public float currentEnergy;
     //public Vector3 cpos, ppos;
