+++ Assets/Scripts/mainSlot.cs
@@ -6,7 +6,7 @@ public class mainSlot : Item {
 	public int healthBoost;
 	public int damage;
 	public int energyBoost;
-	public int atkRate;
+	public float atkRate;
 	public int range;
 
 	// Use this for initialization
